requirements.txt
asgiref==3.2.10
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
Django==3.0.8
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-storages==1.7.1
gcloud==0.18.3
google-api-core==1.22.0
google-auth==1.19.2
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
oauth2client==4.1.3
Pillow==7.2.0
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
django-cors-headers==3.4.0
python-http-client==3.2.7
sendgrid==6.4.5
starkbank-ecdsa==1.0.0
django-sendgrid-v5==0.8.1
future==0.18.2
django-simple-captcha==0.5.12
django-ranged-response==0.2.0



